Separate data processing from API
