+++ server/core/CodeforcesAPI.js
@@ -11,7 +11,7 @@ class CodeforcesAPI {
         });
     }
 
-    static async updateUserStats(username) {
+    static async fetchUserData(username) {
         const user = await prisma.User.findUnique({ where: { username } });
         if (!user || !user.handle) {
             return console.error("No valid user or user handle");
@@ -22,11 +22,7 @@ class CodeforcesAPI {
         } catch (error) {
             return console.error("[Failed to fetch user stats: ", error);
         }
-        await this.processUserSubmissions(data, username);
-        await this.processUserStats(username, user.handle);
-    }
 
-    static async processUserSubmissions(data, username) {
         for (const submission of data.result) {
             // check if submission exists, if it does, that means everything after it also exists
             const existingSubmission = await prisma.Submission.findUnique({
@@ -81,102 +77,7 @@ class CodeforcesAPI {
         }
     }
 
-    static async processUserStats(username) {
-        // count problems AC'ed
-        const totalProblemsAC = await prisma.UserProblemStatus.count({
-            where: {
-                username,
-                AC: { gt: 0 },
-            },
-        });
-
-        // count submission + AC for AC rate, and avg for average rating of problem solved
-        const totalSubmissionsAndAC = await prisma.UserProblemStatus.aggregate({
-            where: { username },
-            _sum: {
-                submissions: true,
-                AC: true,
-            },
-        });
-
-        const problemStatuses = await prisma.UserProblemStatus.findMany({
-            where: { username },
-            include: {
-                problem: { include: { submissions: { orderBy: { timeCreated: "asc" } } } },
-            },
-        });
-
-        // count the frequency of a question tag and tag difficulty
-        const tagsFrequency = {};
-        const tagsDifficulty = {};
-        for (const problemStatus of problemStatuses) {
-            // calculate the difficulty of this problem if user didn't submit
-            // based on submissions before AC (not accurate but gets the job done)
-            let userDifficultyRating = problemStatus.userDifficultyRating;
-            if (userDifficultyRating == -1) {
-                let submissionsBeforeAC = 0;
-                let ACed = false;
-                for (const submission of problemStatus.problem.submissions) {
-                    if (submission.verdict === "OK") {
-                        ACed = true;
-                        break;
-                    }
-                    submissionsBeforeAC++;
-                }
-                userDifficultyRating = ACed ? Math.min(1 + submissionsBeforeAC, 5) : 5;
-                await prisma.UserProblemStatus.update({
-                    where: { username_problemId: { username, problemId: problemStatus.problemId } },
-                    data: { userDifficultyRating },
-                });
-            }
-
-            for (const tag of problemStatus.problem.tags) {
-                if (!tagsFrequency[tag]) {
-                    tagsFrequency[tag] = 0;
-                }
-                tagsFrequency[tag]++;
-                if (!tagsDifficulty[tag]) {
-                    tagsDifficulty[tag] = 0;
-                }
-                tagsDifficulty[tag] += userDifficultyRating;
-            }
-        }
-
-        // count the number of submissions and AC over the last 60 days
-        const past60DaySubmissions = Array(60).fill(0),
-            past60DayAC = Array(60).fill(0);
-        const sortedSubmissions = await prisma.Submission.findMany({
-            where: {
-                authorUsername: username,
-                timeCreated: { gte: new Date(new Date().setDate(new Date().getDate() - 60)) },
-            },
-            orderBy: { timeCreated: "desc" },
-        });
-
-        const timeNow = new Date();
-        for (const submission of sortedSubmissions) {
-            const dayDiff = Math.floor((timeNow - new Date(submission.timeCreated)) / (1000 * 60 * 60 * 24));
-            past60DaySubmissions[dayDiff]++;
-            if (submission.verdict === "OK") {
-                past60DayAC[dayDiff]++;
-            }
-        }
-
-        await prisma.User.update({
-            where: { username },
-            data: {
-                problemsAC: totalProblemsAC,
-                totalSubmissions: totalSubmissionsAndAC._sum.submissions,
-                totalAC: totalSubmissionsAndAC._sum.AC,
-                tagsFrequency,
-                tagsDifficulty,
-                recentSubmissions: past60DaySubmissions,
-                recentAC: past60DayAC,
-            },
-        });
-    }
-
-    static async updateProblems() {
+    static async fetchProblemsData() {
         try {
             const data = await this.get("https://codeforces.com/api/problemset.problems").then((response) => response.json());
             if (data.status != "OK") {
@@ -215,7 +116,8 @@ class CodeforcesAPI {
         }
     }
 
-    static async getUserInfo(handle) {
+    // not the same as user data as this stores general user info like name, not submissions
+    static async fetchUserInfo(handle) {
         try {
             const data = await this.get(`https://codeforces.com/api/user.info?handles=${handle}`).then((response) => response.json());
             return data;
